requirements.txt
numpy
opencv-python
matplotlib
scipy
pillow